Emacs: enable show-trailing-whitespace on per-mode basis, move eldoc cfg
Since we cannot customize fundamental-mode, where we ant keep
show-trailing-whitespace nil, do not enable it globally and try to
disable it in selected modes, but do it the other way around: keep it
disabled and enabled in selected modes.

At the same time move eldoc setup to defaults.el as eldoc is bundled
with Emacs.