Emacs: workaround https://github.com/magit/magit/issues/4098
Workaround https://github.com/magit/magit/issues/4098 - with Magit 2.90.1, C-c
C-d will fail over TRAMP if the repository is actually a worktree. With the
trunk version the commit itself will fail as of 20200627.806. Workaround by
prepending TRAMP prefix to magit-toplevel result if the input directory had it
but the result dropped it.